pyproject.toml
[project]
name = "eizins-nippon-colors"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "html5lib>=1.1",
    "httpx>=0.28.1",
    "lxml>=6.0.0",
    "ruff>=0.12.2",
    "pillow>=10.0.0",
    "black>=25.1.0",
    "tqdm>=4.67.1",
    "jinja2>=3.1.6",
    "click>=8.2.1",
    "cupy-cuda12x>=13.4.1",
    "scikit-image>=0.25.2",
    "colour-science>=0.4.6",
    "matplotlib>=3.10.3",
]
